diarization/audio-only/requirements.txt
boto3==1.21.4
botocore==1.24.5
c==0.0.1
cffi==1.15.0
click==8.0.4
colorama==0.4.4
cycler==0.11.0
docopt==0.6.2
fastcluster==1.2.4
flatbuffers==2.0
fonttools==4.29.1
h5py==3.6.0
intervaltree==3.1.0
jmespath==0.10.0
joblib==1.1.0
kaldi-io==0.9.4
kiwisolver==1.3.2
matplotlib==3.5.1
mpmath==1.2.1
numpy==1.22.2
onnxruntime==1.10.0
packaging==21.3
pandas==1.4.1
Pillow==9.0.1
pkg_resources==0.0.0
protobuf==3.19.4
pyannote.core==4.3
pyannote.database==4.1.3
pyannote.metrics==3.2
pycparser==2.21
pyparsing==3.0.7
python-dateutil==2.8.2
pytz==2021.3
PyYAML==6.0
s3transfer==0.5.1
scikit-learn==1.0.2
scipy==1.8.0
shellingham==1.4.0
simplejson==3.17.6
six==1.16.0
sklearn==0.0
sortedcontainers==2.4.0
SoundFile==0.10.3.post1
sympy==1.9
tabulate==0.8.9
threadpoolctl==3.1.0
torch==1.10.2
tqdm==4.62.3
typer==0.4.0
typing_extensions==4.1.1
urllib3==1.26.8
